8267356: AArch64: Vector API SVE codegen support
This is the integration of current SVE work done in
panama-vector/vectorIntrinscs, which includes:

1. Code generation for Vector API c2 IR nodes with SVE.
2. Non-max vector size support with SVE, e.g. using *128Vector APIs on
   256-bit SVE environment could also generate optimized SVE
   instructions with predicate feature.
3. Some more SVE assemblers (and tests) used by the codegen part.

Test: tier1-3 with vector api test cases passed on 512-bit SVE hardware
with MaxVectorSize=16/32/64.

Note: our original plan was making this part of JEP 414 Vector API
(Second Incubator)[1], but we realized that it's now close to 17 release
cycle and the JEP process may take time. Adding more features may delay
the whole review process for the JEP. So we separate this work out as a
standalone patch.

[1] http://openjdk.java.net/jeps/414